+++ nvim/lua/plugins/lazydev.lua
@@ -15,7 +15,7 @@ return {
     opts = {
       sources = {
         -- add lazydev to your completion providers
-        default = { 'lazydev', 'lsp', 'path', 'snippets', 'buffer' },
+        default = { 'lazydev', 'lsp', 'path', 'snippets', 'buffer', 'emoji', 'minuet' },
         providers = {
           lazydev = {
             name = 'LazyDev',
@@ -23,7 +23,45 @@ return {
             -- make lazydev completions top priority (see `:h blink.cmp`)
             score_offset = 100,
           },
+          emoji = {
+            module = 'blink-emoji',
+            name = 'Emoji',
+            score_offset = -9999, -- Tune by preference
+            opts = { insert = true }, -- Insert emoji (default) or complete its name
+          },
+          minuet = {
+            name = 'minuet',
+            module = 'minuet.blink',
+            score_offset = 100,
+          },
         },
+      },
+      appearance = {
+        nerd_font_variant = 'MyFont Nerd Font Mono',
+      },
+      completion = {
+        menu = {
+          draw = {
+            columns = {
+              { 'kind_icon', 'label', 'label_description', gap = 1 },
+              { 'kind', gap = 1 },
+              { 'source_name' },
+            },
+          },
+        },
+      },
+    },
+  },
+  { 'moyiz/blink-emoji.nvim' },
+  {
+    'folke/lazydev.nvim',
+    ft = 'lua',
+    dependencies = {
+      { 'gonstoll/wezterm-types', lazy = true },
+    },
+    opts = {
+      library = {
+        { path = 'wezterm-types', mods = { 'wezterm' } },
       },
     },
   },