+++ nvim/lua/plugins/modelmate.lua
@@ -4,7 +4,8 @@ return {
     config = function()
       require('nvim-modelmate').setup({
         debug = false,
-        model = 'llama3:8b',
+        model = 'codegemma:latest',
+        -- model = 'llama3:8b',
       })
     end,
   },